% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_df}
\alias{clean_df}
\title{Cleans column names, removes unwanted columns and turns data frame into a tibble.}
\usage{
clean_df(df)
}
\arguments{
\item{df}{Data frame returned by \code{\link[wiesbaden:retrieve_data]{wiesbaden::retrieve_data()}}}
}
\value{
Tibble with cleaned column names and removed columns
}
\description{
Cleans column names, removes unwanted columns and turns data frame into a tibble.
}
\examples{
\dontrun{
df <- wiesbaden::retrieve_data(
  tablename = "21311LS001A",
  genesis = c(db = "nrw")
)

clean_df(df)
}
}
