% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_url.R
\name{col_to_url}
\alias{col_to_url}
\title{Converts a column with URL fragments into the full URL}
\usage{
col_to_url(df, columns)
}
\arguments{
\item{df}{A data frame or tibble}

\item{columns}{Unquoted column names}
}
\value{
A data frame with the columns turned into full URLs
}
\description{
This function relies on \code{gt:::resolve_cols_c} from the gt package
}
\examples{
col_to_url(hochschulen, dplyr::everything())
}
