% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{create_factor}
\alias{create_factor}
\title{Turns character vector into a factor. The labels for the levels are retrieved
by joining \link{value_labels}.}
\usage{
create_factor(x)
}
\arguments{
\item{x}{Character vector containing value labels from the Landesdatenbank NRW}
}
\value{
Factor using the description from \link{value_labels}
}
\description{
Turns character vector into a factor. The labels for the levels are retrieved
by joining \link{value_labels}.
}
\examples{
create_factor(value_labels[["value_label"]])
}
