% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{create_factors}
\alias{create_factors}
\title{Turns all character vectors of the data frame or tibble into vectors using
\code{\link[=create_factor]{create_factor()}}.}
\usage{
create_factors(df)
}
\arguments{
\item{df}{Data frame returned by \code{\link[wiesbaden:retrieve_data]{wiesbaden::retrieve_data()}}}
}
\value{
Data frame with all character vectors turned into factors. If a match
can be found, levels will be retrieved from \link{value_labels}.
}
\description{
Turns all character vectors of the data frame or tibble into vectors using
\code{\link[=create_factor]{create_factor()}}.
}
\examples{
create_factors(value_labels)
}
