% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LJ071.R
\docType{data}
\name{df_21311LJ071}
\alias{df_21311LJ071}
\title{21311LJ071 - Studierende, Nebenhörer, Geschlecht, Hochschule (Standorte), Land, Semester}
\format{
A tibble with 6.772 rows and 6 variables: \describe{ \item{BIL002}{Studierende} \item{BIL045}{Nebenhörer} \item{GES}{Geschlecht} \item{BILHS5}{Hochschule (Standorte)} \item{DLAND}{Land} \item{SEMEST}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2024-05-05) using the code '21311LJ071'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the \href{https://larmarange.github.io/labelled/}{labelled} package. The code for all these transformations is available through this package.
}
\usage{
df_21311LJ071
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LJ071 \cr
   Number of rows \tab 6772 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 4 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   GES \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 2 \tab 0 \cr
   BILHS5 \tab 0 \tab 1 \tab 8 \tab 8 \tab 0 \tab 188 \tab 0 \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
   SEMEST \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 31 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   BIL002 \tab 0 \tab 1 \tab 2678.84 \tab 5460.65 \tab 1 \tab 158.75 \tab 487.5 \tab 1902 \tab 39636 \tab ▇▁▁▁▁ \cr
   BIL045 \tab 0 \tab 1 \tab 44.38 \tab 200.68 \tab 0 \tab 0.00 \tab 0.0 \tab 10 \tab 3007 \tab ▇▁▁▁▁ \cr
}
}
\examples{
library(labelled)

df_21311LJ071

}
\keyword{datasets}
