% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LJ217.R
\docType{data}
\name{df_21311LJ217}
\alias{df_21311LJ217}
\title{21311LJ217 - Ausländische Studierende, Bildungsherkunft, Kontinente, Land, Semester}
\format{
A tibble with 336 rows and 5 variables: \describe{ \item{BIL040}{Ausländische Studierende} \item{BILHER}{Bildungsherkunft} \item{BILKONT}{Kontinente} \item{DLAND}{Land} \item{SEMEST}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2024-05-05) using the code '21311LJ217'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the \href{https://larmarange.github.io/labelled/}{labelled} package. The code for all these transformations is available through this package.
}
\usage{
df_21311LJ217
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LJ217 \cr
   Number of rows \tab 336 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 4 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   BILHER \tab 0 \tab 1 \tab 6 \tab 7 \tab 0 \tab 2 \tab 0 \cr
   BILKONT \tab 0 \tab 1 \tab 3 \tab 6 \tab 0 \tab 6 \tab 0 \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
   SEMEST \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 28 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   BIL040 \tab 0 \tab 1 \tab 5739.37 \tab 8262.3 \tab 6 \tab 175.25 \tab 1198 \tab 7580.25 \tab 37293 \tab ▇▁▁▁▁ \cr
}
}
\examples{
library(labelled)

df_21311LJ217

}
\keyword{datasets}
