% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LJ244.R
\docType{data}
\name{df_21311LJ244}
\alias{df_21311LJ244}
\title{21311LJ244 - Studienanfänger im 1. Hochschulsemester, Studienanfänger im 1. Fachsemester, Geschlecht, Hochschule, Land, Studienjahr}
\format{
A tibble with 3.820 rows and 6 variables: \describe{ \item{BIL041}{Studienanfänger im 1. Hochschulsemester} \item{BIL042}{Studienanfänger im 1. Fachsemester} \item{GES}{Geschlecht} \item{BILHS3}{Hochschule} \item{DLAND}{Land} \item{SJAHR}{Studienjahr} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2024-05-05) using the code '21311LJ244'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the \href{https://larmarange.github.io/labelled/}{labelled} package. The code for all these transformations is available through this package.
}
\usage{
df_21311LJ244
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LJ244 \cr
   Number of rows \tab 3820 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 3 \cr
   numeric \tab 3 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   GES \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 2 \tab 0 \cr
   BILHS3 \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 95 \tab 0 \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   SJAHR \tab 0 \tab 1 \tab 2008.78 \tab 8.53 \tab 1993 \tab 2002 \tab 2010.0 \tab 2016.00 \tab 2022 \tab ▆▆▆▇▇ \cr
   BIL041 \tab 0 \tab 1 \tab 719.93 \tab 931.11 \tab 0 \tab 56 \tab 291.5 \tab 1041.25 \tab 5224 \tab ▇▂▁▁▁ \cr
   BIL042 \tab 0 \tab 1 \tab 1170.82 \tab 1775.19 \tab 0 \tab 79 \tab 371.5 \tab 1459.25 \tab 13033 \tab ▇▁▁▁▁ \cr
}
}
\examples{
library(labelled)

df_21311LJ244

}
\keyword{datasets}
