% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LJ308.R
\docType{data}
\name{df_21311LJ308}
\alias{df_21311LJ308}
\title{21311LJ308 - Studienanfänger im 1. Hochschulsemester, Studienanfänger im 1. Fachsemester, Geschlecht, Fächergruppe, Schulform der Hochschulzugangsberechtigung, Land, Studienjahr}
\format{
A tibble with 3.261 rows and 7 variables: \describe{ \item{BIL041}{Studienanfänger im 1. Hochschulsemester} \item{BIL042}{Studienanfänger im 1. Fachsemester} \item{GES}{Geschlecht} \item{BINWFG3}{Fächergruppe} \item{BILZUBF}{Schulform der Hochschulzugangsberechtigung} \item{DLAND}{Land} \item{SJAHR}{Studienjahr} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2024-05-05) using the code '21311LJ308'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: Compared to the raw data, the following changes were made: (1) technical columns ending in "QUAL", "ERR", or "LOCK" were removed, and (2) variable and value labels were added to each table using the \href{https://larmarange.github.io/labelled/}{labelled} package. The code for all these transformations is available through this package.
}
\usage{
df_21311LJ308
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LJ308 \cr
   Number of rows \tab 3261 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 4 \cr
   numeric \tab 3 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   GES \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 2 \tab 0 \cr
   BINWFG3 \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 8 \tab 0 \cr
   BILZUBF \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 7 \tab 0 \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   SJAHR \tab 0 \tab 1 \tab 2007.79 \tab 8.58 \tab 1993 \tab 2000 \tab 2008 \tab 2015 \tab 2022 \tab ▇▇▇▇▇ \cr
   BIL041 \tab 0 \tab 1 \tab 843.34 \tab 1805.73 \tab 0 \tab 39 \tab 179 \tab 714 \tab 16937 \tab ▇▁▁▁▁ \cr
   BIL042 \tab 0 \tab 1 \tab 1371.52 \tab 3145.12 \tab 0 \tab 60 \tab 308 \tab 1094 \tab 30115 \tab ▇▁▁▁▁ \cr
}
}
\examples{
library(labelled)

df_21311LJ308

}
\keyword{datasets}
