% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LS001A.R
\docType{data}
\name{df_21311LS001A}
\alias{df_21311LS001A}
\title{21311LS001A - Studierende / Studenten, Land, Hochschulen, Studienfach, Fachsemester, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 653.262 rows and 7 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{bilfs1}{Fachsemester}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311LS001A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LS001A \cr
   Number of rows \tab 653262 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 6 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 653262 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 61657, Uni: 59914, Uni: 50567, Uni: 38768 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 234 \tab Ger: 25322, Mat: 23709, Wir: 22203, Ang: 20419 \cr
   bilfs1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 57021, 3. : 51681, 5. : 48343, 7. : 44034 \cr
   bilap1 \tab 0 \tab 1 \tab FALSE \tab 59 \tab Pro: 106406, Bac: 59918, Mas: 47424, Dip: 46719 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 39053, WS : 38035, WS : 36760, WS : 35569 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 19.96 \tab 55.9 \tab 1 \tab 2 \tab 4 \tab 16 \tab 3843 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311LS001A
}
\keyword{datasets}
