% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LS002A.R
\docType{data}
\name{df_21311LS002A}
\alias{df_21311LS002A}
\title{21311LS002A - Studierende / Studenten, Land, Hochschulen, Studienfach, Hochschulsemester, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 801.357 rows and 7 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{bilsh1}{Hochschulsemester}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311LS002A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LS002A \cr
   Number of rows \tab 801357 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 6 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 801357 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 69710, Uni: 69128, Uni: 53576, Uni: 45978 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 234 \tab Ger: 29102, Wir: 28897, Mat: 28025, Che: 24156 \cr
   bilsh1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 20.: 58995, 13.: 49735, 11.: 49537, 9. : 48102 \cr
   bilap1 \tab 0 \tab 1 \tab FALSE \tab 59 \tab Bac: 97664, Pro: 90045, Mas: 71913, Bac: 68888 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 48361, WS : 47698, WS : 46596, WS : 45738 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 16.27 \tab 48.03 \tab 1 \tab 1 \tab 4 \tab 12 \tab 4181 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311LS002A
}
\keyword{datasets}
