% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW001A.R
\docType{data}
\name{df_21311NW001A}
\alias{df_21311NW001A}
\title{21311NW001A - Studierende / Studenten, Land, Hochschulen, Fachsemester, Geschlecht, Hörerstatus, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 239.664 rows and 8 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilfs1}{Fachsemester}
\item{ges}{Geschlecht}
\item{bilhn1}{Hörerstatus}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW001A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW001A \cr
   Number of rows \tab 239664 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 7 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 239664 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 15140, Uni: 14905, Uni: 12164, Tec: 10890 \cr
   bilfs1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 20315, 3. : 18801, 5. : 17675, 7. : 16003 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 122221, wei: 117443 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 200113, Neb: 39551 \cr
   bilap1 \tab 0 \tab 1 \tab FALSE \tab 59 \tab Bac: 27553, Dip: 21997, Pro: 17090, Dip: 14374 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 13009, WS : 12862, WS : 12436, WS : 12023 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 54.42 \tab 162.34 \tab 1 \tab 2 \tab 8 \tab 37 \tab 6840 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW001A
}
\keyword{datasets}
