% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW001B.R
\docType{data}
\name{df_21311NW001B}
\alias{df_21311NW001B}
\title{21311NW001B - Studierende / Studenten, Land, Hochschulen, Fachsemester, Nationalität, Hörerstatus, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 214.109 rows and 8 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilfs1}{Fachsemester}
\item{nat}{Nationalität}
\item{bilhn1}{Hörerstatus}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
}
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-11) using the code 21311NW001B. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311NW001B
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW001B \cr
   Number of rows \tab 214109 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 7 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 214109 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 13377, Uni: 12722, Uni: 11084, Tec: 9301 \cr
   bilfs1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 18335, 3. : 16867, 5. : 15835, 7. : 14371 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Deu: 133414, Aus: 80695 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 180817, Neb: 33292 \cr
   bilap1 \tab 0 \tab 1 \tab FALSE \tab 59 \tab Bac: 25225, Dip: 20206, Pro: 16443, Dip: 13514 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 11675, WS : 11581, WS : 11114, WS : 10778 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 60.91 \tab 213.05 \tab 1 \tab 2 \tab 7 \tab 34 \tab 11242 \tab ▇▁▁▁▁ \cr
}
}
\examples{
df_21311NW001B
}
\keyword{datasets}
