% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW002B.R
\docType{data}
\name{df_21311NW002B}
\alias{df_21311NW002B}
\title{21311NW002B - Studierende / Studenten, Land, Hochschulen, Fachsemester, Nationalität, Semester}
\format{
A tibble with 58.698 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilfs1}{Fachsemester}
\item{nat}{Nationalität}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW002B
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW002B \cr
   Number of rows \tab 58698 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 58698 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 880, Uni: 880, Uni: 880, Fer: 880 \cr
   bilfs1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 4345, 3. : 4251, 5. : 4147, 7. : 3888 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Deu: 33274, Aus: 25424 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 3596, WS : 3531, WS : 3348, WS : 3260 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 222.18 \tab 674.72 \tab 1 \tab 5 \tab 26 \tab 137 \tab 14294 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW002B
}
\keyword{datasets}
