% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW003A.R
\docType{data}
\name{df_21311NW003A}
\alias{df_21311NW003A}
\title{21311NW003A - Studierende / Studenten, Land, Hochschulen, Geschlecht, Hörerstatus, Semester}
\format{
A tibble with 7.137 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{ges}{Geschlecht}
\item{bilhn1}{Hörerstatus}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW003A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW003A \cr
   Number of rows \tab 7137 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 7137 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 88, Uni: 88, Fer: 88, Uni: 88 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 3614, wei: 3523 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 4888, Neb: 2249 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 418, WS : 417, WS : 402, WS : 388 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 1827.34 \tab 4621.62 \tab 1 \tab 29 \tab 213 \tab 891 \tab 39636 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW003A
}
\keyword{datasets}
