% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW003B.R
\docType{data}
\name{df_21311NW003B}
\alias{df_21311NW003B}
\title{21311NW003B - Studierende / Studenten, Land, Hochschulen, Nationalität, Hörerstatus, Semester}
\format{
A tibble with 6.690 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{nat}{Nationalität}
\item{bilhn1}{Hörerstatus}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW003B
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW003B \cr
   Number of rows \tab 6690 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 6690 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 88, Uni: 88, Fer: 88, Uni: 88 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Deu: 3684, Aus: 3006 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 4579, Neb: 2111 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 398, WS : 392, WS : 375, WS : 366 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 1949.43 \tab 6087.9 \tab 1 \tab 13 \tab 132 \tab 967.75 \tab 68398 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW003B
}
\keyword{datasets}
