% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW004A.R
\docType{data}
\name{df_21311NW004A}
\alias{df_21311NW004A}
\title{21311NW004A - Studierende / Studenten, Land, Studienfach, Geschlecht, Hörerstatus, Semester}
\format{
A tibble with 13.451 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilsf1}{Studienfach}
\item{ges}{Geschlecht}
\item{bilhn1}{Hörerstatus}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW004A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW004A \cr
   Number of rows \tab 13451 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 13451 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 234 \tab Ang: 88, Arc: 88, Bau: 88, Bet: 88 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 6768, wei: 6683 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 8403, Neb: 5048 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 646, WS : 646, WS : 644, WS : 644 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 969.57 \tab 2994.3 \tab 1 \tab 8 \tab 86 \tab 496 \tab 40048 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW004A
}
\keyword{datasets}
