% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW005A.R
\docType{data}
\name{df_21311NW005A}
\alias{df_21311NW005A}
\title{21311NW005A - Studierende / Studenten, Land, Hochschulen, Hochschulsemester, Geschlecht, Hörerstatus, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 292.667 rows and 8 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilsh1}{Hochschulsemester}
\item{ges}{Geschlecht}
\item{bilhn1}{Hörerstatus}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW005A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW005A \cr
   Number of rows \tab 292667 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 7 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 292667 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 17720, Uni: 17056, Uni: 12885, Tec: 12799 \cr
   bilsh1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 20.: 19898, 11.: 17626, 9. : 17535, 7. : 17228 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 148832, wei: 143835 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 248922, Neb: 43745 \cr
   bilap1 \tab 0 \tab 1 \tab FALSE \tab 59 \tab Bac: 39422, Dip: 25591, Mas: 23584, Pro: 16326 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 15867, WS : 15809, WS : 15447, WS : 15151 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 44.56 \tab 138.88 \tab 1 \tab 2 \tab 6 \tab 28 \tab 9244 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW005A
}
\keyword{datasets}
