% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW006A.R
\docType{data}
\name{df_21311NW006A}
\alias{df_21311NW006A}
\title{21311NW006A - Studierende / Studenten, Land, Hochschulen, Studienfach, Hochschulsemester, Geschlecht, Semester}
\format{
A tibble with 701.621 rows and 7 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{bilsh1}{Hochschulsemester}
\item{ges}{Geschlecht}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW006A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW006A \cr
   Number of rows \tab 701621 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 6 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 701621 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 51163, Uni: 45237, Uni: 42470, Uni: 38347 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 234 \tab Wir: 28646, Bet: 24943, Mas: 18955, Ele: 17956 \cr
   bilsh1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 42939, 7. : 42641, 3. : 42504, 5. : 42135 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 359177, wei: 342444 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 43076, WS : 42199, WS : 40732, WS : 39249 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 18.59 \tab 49.01 \tab 1 \tab 2 \tab 5 \tab 17 \tab 4798 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW006A
}
\keyword{datasets}
