% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW008.R
\docType{data}
\name{df_21311NW008}
\alias{df_21311NW008}
\title{21311NW008 - Studierende / Studenten, Land, Nationalität, Geschlecht, Hörerstatus, Semester}
\format{
A tibble with 264 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{nat}{Nationalität}
\item{ges}{Geschlecht}
\item{bilhn1}{Hörerstatus}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW008
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW008 \cr
   Number of rows \tab 264 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 264 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Aus: 132, Deu: 132 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 132, wei: 132 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 176, Neb: 88 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 44 \tab WS : 8, WS : 8, WS : 8, WS : 8 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 84400.77 \tab 108026.1 \tab 127 \tab 3785 \tab 25683 \tab 182503.2 \tab 354416 \tab ▇▁▂▂▁ \cr
}


NULL
}
\examples{
df_21311NW008
}
\keyword{datasets}
