% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW009.R
\docType{data}
\name{df_21311NW009}
\alias{df_21311NW009}
\title{21311NW009 - Studierende / Studenten, Land, Nationalität, Geschlecht, Semester}
\format{
A tibble with 176 rows and 5 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{nat}{Nationalität}
\item{ges}{Geschlecht}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW009
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW009 \cr
   Number of rows \tab 176 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 4 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 176 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Aus: 88, Deu: 88 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 88, wei: 88 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 44 \tab WS : 4, WS : 4, WS : 4, WS : 4 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 126601.1 \tab 112722.4 \tab 3571 \tab 25911.5 \tab 71710.5 \tab 222994 \tab 358343 \tab ▇▁▃▃▂ \cr
}


NULL
}
\examples{
df_21311NW009
}
\keyword{datasets}
