% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW010.R
\docType{data}
\name{df_21311NW010}
\alias{df_21311NW010}
\title{21311NW010 - Studierende / Studenten, Land, Nationalität, Hörerstatus, Semester}
\format{
A tibble with 132 rows and 5 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{nat}{Nationalität}
\item{bilhn1}{Hörerstatus}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW010
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW010 \cr
   Number of rows \tab 132 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 4 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 132 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Aus: 66, Deu: 66 \cr
   bilhn1 \tab 0 \tab 1 \tab FALSE \tab 2 \tab Hau: 88, Neb: 44 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 44 \tab WS : 4, WS : 4, WS : 4, WS : 4 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 168801.5 \tab 212651.6 \tab 333 \tab 7537.5 \tab 47243 \tab 397062.8 \tab 676226 \tab ▇▁▁▂▁ \cr
}


NULL
}
\examples{
df_21311NW010
}
\keyword{datasets}
