% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW013.R
\docType{data}
\name{df_21311NW013}
\alias{df_21311NW013}
\title{21311NW013 - Studierende / Studenten, Land, Nationalität, Semester}
\format{
A tibble with 44 rows and 4 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{nat}{Nationalität}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW013
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW013 \cr
   Number of rows \tab 44 \cr
   Number of columns \tab 4 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 3 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 44 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Aus: 22, Deu: 22 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 2, WS : 2, WS : 2, WS : 2 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 296402.2 \tab 241190.6 \tab 49717 \tab 63023.25 \tab 252554 \tab 465787 \tab 684797 \tab ▇▁▂▃▃ \cr
}


NULL
}
\examples{
df_21311NW013
}
\keyword{datasets}
