% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW014.R
\docType{data}
\name{df_21311NW014}
\alias{df_21311NW014}
\title{21311NW014 - Studienanfänger, Geschlecht, Hochschulen, Studienfach, Land, Semester}
\format{
A tibble with:
\describe{
\item{bil016}{Studienanfänger}
\item{ges}{Geschlecht}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{dland}{Land}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-10
}
\usage{
df_21311NW014
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\strong{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\strong{Education and Culture}) -> 213 Hochschulen (\strong{Universities}) -> 21311 Statistik der Studenten (\strong{Student Statistics}).
}
\keyword{datasets}
