% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Generate roxygen documentation for a table code}
\format{
A tibble with:
\\describe{",
}
\source{
\url{https://www.landesdatenbank.nrw.de}, {{date}}
}
\usage{
document_data(
  table_name,
  table_description,
  table_rda,
  download_date = Sys.Date()
)
}
\arguments{
\item{table_name}{Table name as a string}

\item{table_description}{Table description as a string}

\item{table_rda}{Name of the rda file as string}

\item{download_date}{Date on which the data was retrieved, defaults to \code{Sys.Date()}}
}
\value{
The roxygen documentation as character
}
\description{
Generate roxygen documentation for a table code
}
\examples{

The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (**Education, Social Benefits, Health, Administration of Justice**) -> 21 Bildung und Kultur (**Education and Culture**) -> 213 Hochschulen (**Universities**) -> 21311 Statistik der Studenten (**Student Statistics**).

}
