% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_format_section}
\alias{document_format_section}
\title{Get roxygen format section for a given table name}
\usage{
document_format_section(table_name)
}
\arguments{
\item{table_name}{Table name as character}
}
\value{
Character vector containing the roxygen format section
}
\description{
Get roxygen format section for a given table name
}
