% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_items}
\alias{document_items}
\title{Gets the item list for each table used in the roxygen documentation}
\usage{
document_items(table_name)
}
\arguments{
\item{table_name}{Table name}
}
\value{
Character vector containing the roxygen item list to be used in the format/describe section
}
\description{
Gets the item list for each table used in the roxygen documentation
}
