% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_table}
\alias{document_table}
\title{Get text for the R file documenting a table}
\usage{
document_table(table_name)
}
\arguments{
\item{table_name}{Valid table name existing in \link{tables}}
}
\value{
Character vector for the r file creating the roxygen dataset documentation
}
\description{
Get text for the R file documenting a table
}
\examples{
document_table(table_name = tables[[1, "tablename"]])
}
