% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_title}
\alias{document_title}
\title{Get title for roxygen data table documentation}
\usage{
document_title(table_name)
}
\arguments{
\item{table_name}{String with table name or code}
}
\value{
Title as character (Table code - Table description)
}
\description{
Get title for roxygen data table documentation
}
\examples{
document_title("21311LS001A")
document_title("df_21311LS001A")
}
