% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{generate_items_roxy}
\alias{generate_items_roxy}
\title{Generates the item list for a table code}
\usage{
generate_items_roxy(table_name)
}
\arguments{
\item{table_name}{Table name as a string}
}
\value{
Roxygen item list
}
\description{
Generates the item list for a table code
}
