% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_ncol}
\alias{get_ncol}
\title{Get formatted number of columns / variables in a data frame}
\usage{
get_ncol(x)
}
\arguments{
\item{x}{df}
}
\value{
Formatted string with number of columns
}
\description{
Get formatted number of columns / variables in a data frame
}
\examples{
studentenstatistikNRW:::get_ncol(mtcars)
}
