% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_nrow}
\alias{get_nrow}
\title{Get formatted number of rows in a data frame}
\usage{
get_nrow(x)
}
\arguments{
\item{x}{df}
}
\value{
Formatted string with number of rows
}
\description{
Get formatted number of rows in a data frame
}
