% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_table_name}
\alias{get_table_name}
\title{Takes table code or name as input and returns table name or code}
\usage{
get_table_name(table_name, prefix = TRUE)
}
\arguments{
\item{table_name}{String with table name or code}

\item{prefix}{Boolean, should result include "df_" prefix or not}
}
\value{
Table name (e.g. df_21311LJ139), or code (e.g. 21311LJ139)
}
\description{
Takes table code or name as input and returns table name or code
}
\examples{
get_table_name(tables[[1, "tablename"]])
get_table_name(tables[[1, "tablename"]], prefix = FALSE)
}
