% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hochschulen.R
\docType{data}
\name{hochschulen}
\alias{hochschulen}
\title{Hochschulen - Higher education institutions in Germany}
\format{
A tibble with 540 rows and 16 variables:
\describe{
\item{Hochschule_Signatur}{3-digit Destatis ID}
}
}
\source{
TO DO: Source ergänzen
}
\usage{
hochschulen
}
\description{
TO DO: Datensatzbeschreibung
}
\examples{
hochschulen

}
\keyword{datasets}
