% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_values}
\alias{label_values}
\title{Set value labels using \link[labelled:val_labels]{labelled::set_value_labels}}
\usage{
label_values(df, table_name)
}
\arguments{
\item{df}{Data frame}

\item{table_name}{String with table name}
}
\value{
Data frame with labeled values
}
\description{
Set value labels using \link[labelled:val_labels]{labelled::set_value_labels}
}
\examples{
library(labelled)

label_values(
  df = df_21311LJ001,
  table_name = "21311LJ001"
)
}
