% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_variables}
\alias{label_variables}
\title{Set variable labels using \link[labelled:var_label]{labelled::set_variable_labels}}
\usage{
label_variables(df, table_name)
}
\arguments{
\item{df}{Data frame}

\item{table_name}{String with table name}
}
\value{
Data frame with labeled variables
}
\description{
Set variable labels using \link[labelled:var_label]{labelled::set_variable_labels}
}
\examples{
label_variables(
  df_21311LS001A,
  table_name = "21311LS001A"
)
}
