% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata containing the variables for each student statistics table}
\format{
A tibble with 1.881 rows and 4 variables:
\describe{
\item{tablename}{Table name}
\item{name}{Variable name}
\item{description}{Variable description}
\item{unit}{Unit for each measure}
}
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-11). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.
}
\usage{
metadata
}
\description{
The metadata for each table name in \link{tables} is sourced using \code{\link[wiesbaden:retrieve_metadata]{wiesbaden::retrieve_metadata()}}. Data source is the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW}.
}
\examples{
metadata

}
\keyword{datasets}
