% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentenstatistikNRW-package.R
\docType{package}
\name{studentenstatistikNRW-package}
\alias{studentenstatistikNRW}
\alias{studentenstatistikNRW-package}
\title{studentenstatistikNRW: Student Statistics from the Landesdatenbank NRW (Table Series 21311)}
\description{
Student statistics for the German state North Rhine-Westphalia obtained from the Landesdatenbank NRW. Packaged for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RichardMeyer-Eppler/studentenstatistikNRW}
  \item \url{https://richardmeyer-eppler.github.io/studentenstatistikNRW/}
  \item Report bugs at \url{https://github.com/RichardMeyer-Eppler/studentenstatistikNRW/issues}
}

}
\author{
\strong{Maintainer}: Richard Meyer-Eppler \email{richard.meyereppler@gmail.com}

}
\keyword{internal}
