% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_info.R
\docType{data}
\name{variable_info}
\alias{variable_info}
\title{Additional information and comments for some variables}
\format{
A tibble with 16 rows and 2 variables:
\describe{
\item{variable}{Variable name}
\item{description}{Additional information for each variable}
}
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2024-05-05). The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.
}
\usage{
variable_info
}
\description{
The additional information for each variable is sourced using \code{\link[wiesbaden:retrieve_varinfo]{wiesbaden::retrieve_varinfo()}}. Data source is the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW}.
}
\examples{
variable_info

}
\keyword{datasets}
