% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_df}
\alias{clean_df}
\title{Cleans column names, removes unwanted columns and turns data frame into a
tibble.}
\usage{
clean_df(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Cleans column names, removes unwanted columns and turns data frame into a
tibble.
}
\examples{
\dontrun{
clean_df(df)
}
}
