% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{create_factor}
\alias{create_factor}
\title{Turns character vector into a factor. The labels for the levels are retrieved
by joining \code{df_value_labels}.}
\usage{
create_factor(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Turns character vector into a factor. The labels for the levels are retrieved
by joining \code{df_value_labels}.
}
\examples{
\dontrun{
create_factor(x)
}
}
