% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{create_factors}
\alias{create_factors}
\title{Turns all character vectors of the data frame or tibble into vectors using
\code{create_factor}.}
\usage{
create_factors(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Turns all character vectors of the data frame or tibble into vectors using
\code{create_factor}.
}
\examples{
\dontrun{
create_factors(df)
}
}
