% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311LS002A.R
\docType{data}
\name{df_21311LS002A}
\alias{df_21311LS002A}
\title{21311LS002A - Studierende / Studenten, Land, Hochschulen, Studienfach, Hochschulsemester, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with 801.357 rows and 7 variables: \describe{ \item{bil002}{Studierende / Studenten} \item{dland}{Land} \item{bilhs1}{Hochschulen} \item{bilsf1}{Studienfach} \item{bilsh1}{Hochschulsemester} \item{bilap1}{Angestrebte Abschlussprüfung} \item{semest}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-12) using the code '21311LS002A'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311LS002A
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311LS002A \cr
   Number of rows \tab 801357 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
   BILHS1 \tab 0 \tab 1 \tab 6 \tab 6 \tab 0 \tab 174 \tab 0 \cr
   BILSF1 \tab 0 \tab 1 \tab 5 \tab 5 \tab 0 \tab 234 \tab 0 \cr
   BILSH1 \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 20 \tab 0 \cr
   BILAP1 \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 59 \tab 0 \cr
   SEMEST \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 22 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   BIL002 \tab 0 \tab 1 \tab 16.27 \tab 48.03 \tab 1 \tab 1 \tab 4 \tab 12 \tab 4181 \tab ▇▁▁▁▁ \cr
}
}
\examples{
df_21311LS002A

}
\keyword{datasets}
