% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW002A.R
\docType{data}
\name{df_21311NW002A}
\alias{df_21311NW002A}
\title{21311NW002A - Studierende / Studenten, Land, Hochschulen, Fachsemester, Geschlecht, Semester}
\format{
A tibble with 63.000 rows and 6 variables:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilfs1}{Fachsemester}
\item{ges}{Geschlecht}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW002A
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW002A \cr
   Number of rows \tab 63000 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 63000 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 880, Uni: 880, Uni: 880, Fer: 880 \cr
   bilfs1 \tab 0 \tab 1 \tab FALSE \tab 20 \tab 1. : 4711, 3. : 4595, 5. : 4499, 7. : 4187 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 32354, wei: 30646 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 3918, WS : 3826, WS : 3651, WS : 3531 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 207.01 \tab 515.21 \tab 1 \tab 6 \tab 33 \tab 158 \tab 8335 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW002A
}
\keyword{datasets}
