% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW002B.R
\docType{data}
\name{df_21311NW002B}
\alias{df_21311NW002B}
\title{21311NW002B - Studierende / Studenten, Land, Hochschulen, Fachsemester, Nationalität, Semester}
\format{
A tibble with:
\describe{
\item{bil002}{Studierende / Studenten}
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilfs1}{Fachsemester}
\item{nat}{Nationalität}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-10
}
\usage{
df_21311NW002B
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\strong{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\strong{Education and Culture}) -> 213 Hochschulen (\strong{Universities}) -> 21311 Statistik der Studenten (\strong{Student Statistics}).
}
\keyword{datasets}
