% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW002C.R
\docType{data}
\name{df_21311NW002C}
\alias{df_21311NW002C}
\title{21311NW002C - Studierende / Studenten, Land, Fachsemester, Nationalität, Geschlecht, Semester}
\format{
A tibble with 1.760 rows and 6 variables: \describe{ \item{bil002}{Studierende / Studenten} \item{dland}{Land} \item{bilfs1}{Fachsemester} \item{nat}{Nationalität} \item{ges}{Geschlecht} \item{semest}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-12) using the code '21311NW002C'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311NW002C
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW002C \cr
   Number of rows \tab 1760 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   DLAND \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 1 \tab 0 \cr
   BILFS1 \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 20 \tab 0 \cr
   NAT \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 2 \tab 0 \cr
   GES \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 2 \tab 0 \cr
   SEMEST \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 22 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   BIL002 \tab 0 \tab 1 \tab 7410.05 \tab 12350.84 \tab 79 \tab 912 \tab 2608 \tab 7497.75 \tab 78016 \tab ▇▁▁▁▁ \cr
}
}
\examples{
df_21311NW002C

}
\keyword{datasets}
