% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW007A.R
\docType{data}
\name{df_21311NW007A}
\alias{df_21311NW007A}
\title{21311NW007A - Studierende / Studenten, Land, Hochschulen, Studienfach, Geschlecht, Semester}
\format{
A tibble with 53.814 rows and 6 variables: \describe{ \item{bil002}{Studierende / Studenten} \item{dland}{Land} \item{bilhs1}{Hochschulen} \item{bilsf1}{Studienfach} \item{ges}{Geschlecht} \item{semest}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-12) using the code '21311NW007A'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311NW007A
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW007A \cr
   Number of rows \tab 53814 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 53814 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 172 \tab Uni: 3341, Uni: 3244, Uni: 3012, Uni: 2720 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 234 \tab Wir: 1968, Bet: 1756, Mas: 1323, Ele: 1276 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 26953, wei: 26861 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 3356, WS : 3260, WS : 3162, WS : 3054 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 242.35 \tab 519.73 \tab 1 \tab 24 \tab 89 \tab 266 \tab 16714 \tab ▇▁▁▁▁ \cr
}
}
\examples{
df_21311NW007A

}
\keyword{datasets}
