% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW009.R
\docType{data}
\name{df_21311NW009}
\alias{df_21311NW009}
\title{21311NW009 - Studierende / Studenten, Land, Nationalität, Geschlecht, Semester}
\format{
A tibble with 176 rows and 5 variables: \describe{ \item{bil002}{Studierende / Studenten} \item{dland}{Land} \item{nat}{Nationalität} \item{ges}{Geschlecht} \item{semest}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-12) using the code '21311NW009'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311NW009
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW009 \cr
   Number of rows \tab 176 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 4 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 176 \cr
   nat \tab 0 \tab 1 \tab FALSE \tab 2 \tab Aus: 88, Deu: 88 \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab män: 88, wei: 88 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 44 \tab WS : 4, WS : 4, WS : 4, WS : 4 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil002_val \tab 0 \tab 1 \tab 126601.1 \tab 112722.4 \tab 3571 \tab 25911.5 \tab 71710.5 \tab 222994 \tab 358343 \tab ▇▁▃▃▂ \cr
}
}
\examples{
df_21311NW009

}
\keyword{datasets}
