% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW014.R
\docType{data}
\name{df_21311NW014}
\alias{df_21311NW014}
\title{21311NW014 - Studienanfänger, Geschlecht, Hochschulen, Studienfach, Land, Semester}
\format{
A tibble with 42.939 rows and 6 variables:
\describe{
\item{bil016}{Studienanfänger}
\item{ges}{Geschlecht}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{dland}{Land}
\item{semest}{Semester}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, 2021-07-11
}
\usage{
df_21311NW014
}
\description{
The table is sourced from the Landesdatenbank NRW from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW014 \cr
   Number of rows \tab 42939 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab wei: 21683, män: 21256 \cr
   bilhs1 \tab 0 \tab 1 \tab FALSE \tab 167 \tab Uni: 2668, Uni: 2455, Uni: 2275, Uni: 2056 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 233 \tab Wir: 1619, Bet: 1584, Mas: 1225, Ele: 1199 \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 42939 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 2621, WS : 2557, WS : 2530, WS : 2454 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil016_val \tab 0 \tab 1 \tab 42.22 \tab 66.83 \tab 1 \tab 7 \tab 20 \tab 52 \tab 1988 \tab ▇▁▁▁▁ \cr
}


NULL
}
\examples{
df_21311NW014
}
\keyword{datasets}
