% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_21311NW015.R
\docType{data}
\name{df_21311NW015}
\alias{df_21311NW015}
\title{21311NW015 - Studienanfänger, Geschlecht, Studienfach, Land, Semester}
\format{
A tibble with 7.579 rows and 5 variables: \describe{ \item{bil016}{Studienanfänger} \item{ges}{Geschlecht} \item{bilsf1}{Studienfach} \item{dland}{Land} \item{semest}{Semester} }
}
\source{
The data in this package was obtained from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} (retrieved on 2021-07-12) using the code '21311NW015'. The licence for data obtained from the Landesdatenbank NRW / Destatis is available here (see paragraph 2 section 2): \href{http://www.govdata.de/dl-de/by-2-0}{Data licence Germany – attribution – Version 2.0}. Paragraph 2 section 3 requires including a reference to the dataset (URI) if available. To the best of my knowledge, the Landesdatenbank NRW does not yet provide Uniform Resource Identifiers (URI) to identify their tables.

Paragraph 3 stipulates that 'changes, editing, new designs or other amendments must be marked as such in the source note'. Compared to the raw data, the following changes were made: (1) column names were changed to lower case, (2) some technical columns were excluded and (3) value labels were turned into factors using the description of each value label. The code for all these transformations is available through this package.
}
\usage{
df_21311NW015
}
\description{
The table is sourced from the \href{https://www.landesdatenbank.nrw.de}{Landesdatenbank NRW} from the series 2 Bildung, Sozialleistungen, Gesundheit, Rechtspflege (\emph{Education, Social Benefits, Health, Administration of Justice}) -> 21 Bildung und Kultur (\emph{Education and Culture}) -> 213 Hochschulen (\emph{Universities}) -> 21311 Statistik der Studenten (\emph{Student Statistics}).
}
\details{
The following summary is automatically generated using \code{\link[skimr:skim]{skimr::skim()}}
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_21311NW015 \cr
   Number of rows \tab 7579 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   factor \tab 4 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   ges \tab 0 \tab 1 \tab FALSE \tab 2 \tab wei: 3808, män: 3771 \cr
   bilsf1 \tab 0 \tab 1 \tab FALSE \tab 233 \tab Agr: 44, All: 44, All: 44, Ang: 44 \cr
   dland \tab 0 \tab 1 \tab FALSE \tab 1 \tab Nor: 7579 \cr
   semest \tab 0 \tab 1 \tab FALSE \tab 22 \tab WS : 378, WS : 376, WS : 375, WS : 371 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   bil016_val \tab 0 \tab 1 \tab 239.22 \tab 541.14 \tab 1 \tab 15 \tab 54 \tab 198.5 \tab 6083 \tab ▇▁▁▁▁ \cr
}
}
\examples{
df_21311NW015

}
\keyword{datasets}
