% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_21311LS001A.R
\docType{data}
\name{table_21311LS001A}
\alias{table_21311LS001A}
\title{Studierende / Studenten, Land, Hochschulen, Studienfach, Fachsemester, Angestrebte Abschlussprüfung, Semester}
\format{
A tibble with columns:
\describe{
\item{dland}{Land}
\item{bilhs1}{Hochschulen}
\item{bilsf1}{Studienfach}
\item{bilfs1}{Fachsemester}
\item{bilap1}{Angestrebte Abschlussprüfung}
\item{semest}{Semester}
\item{bil002_val}{Studierende / Studenten (Anzahl)}
}
}
\source{
\url{https://www.landesdatenbank.nrw.de}, downloaded 2021-07-09
}
\usage{
table_21311LS001A
}
\description{
Blah...
}
\examples{
table_21311LS001A
}
\keyword{datasets}
